# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API SmartQuote",
    "description": "API para gerenciamento do SmartQuote.",
    "version": "1.0.0"
  },
  "host": "https://testsmart-24vt.onrender.com",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/produtos/{id}": {
      "get": {
        "summary": "Obter produto por ID",
        "description": "Retorna um produto específico com base no ID fornecido.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do produto",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Produto encontrado"
          },
          "404": {
            "description": "Produto não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "patch": {
        "summary": "Atualizar parcialmente um produto",
        "description": "Atualiza um ou mais campos de um produto com base no ID fornecido.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do produto a ser atualizado",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Campos do produto a serem atualizados",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProdutoUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto atualizado com sucesso"
          },
          "404": {
            "description": "Produto não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "delete": {
        "summary": "Deletar um produto",
        "description": "Deleta um produto do sistema com base no ID fornecido.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do produto a ser deletado",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Produto deletado com sucesso"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ProdutoUpdate": {
      "type": "object",
      "properties": {
        "fornecedor_id": { "type": "integer", "example": 1 },
        "nome": { "type": "string", "example": "Novo Notebook" },
        "preco": { "type": "number", "example": 400000 },
        "estoque": { "type": "integer", "example": 15 },
        "descricao": { "type": "string", "example": "Notebook com 32GB RAM" },
        "codigo": { "type": "string", "example": "F1-NOTE-123456" },
        "modelo": { "type": "string", "example": "i7 16GB 512GB" },
        "origem": { "type": "string", "enum": ["local", "externo"], "example": "externo" },
        "image_url": { "type": "string", "example": "https://example.com/img.jpg" },
        "produto_url": { "type": "string", "example": "https://loja.com/produto" },
        "tags": { "type": "array", "items": {"type": "string"}, "example": ["notebook","intel"] },
        "categoria": { "type": "string", "example": "Informática" },
        "disponibilidade": { "type": "string", "enum": ["imediata","por encomenda","limitada"], "example": "imediata" },
        "especificacoes_tecnicas": { "type": "object", "example": {"tela":"15.6","ram":"16GB"} }
      }
    },
...
  }
}